+++ tests/Basket/BasketServiceTest.php
@@ -10,8 +10,6 @@ namespace Wizaplace\SDK\Tests\Basket;
 use Wizaplace\SDK\Basket\BasketComment;
 use Wizaplace\SDK\Basket\BasketItems;
 use Wizaplace\SDK\Basket\BasketService;
-use Wizaplace\SDK\Basket\Payment;
-use Wizaplace\SDK\Basket\PaymentType;
 use Wizaplace\SDK\Basket\ProductComment;
 use Wizaplace\SDK\Basket\Shipping;
 use Wizaplace\SDK\Catalog\DeclinationId;
@@ -645,6 +643,22 @@ final class BasketServiceTest extends ApiTestCase
 
         $basket = $service->getBasket($basket->getId());
         static::assertSame(10.0, $basket->getTotalMarketplaceDiscount());
+    }
+
+    public function testDeleteUserBasket(): void
+    {
+        $basketId = '8c54d443-eef1-4086-9bb7-b8917d8710e3';
+
+        $apiClient = $this->buildApiClient();
+        $apiClient->authenticate('user@wizaplace.com', 'password');
+        $service = new BasketService($apiClient);
+
+        $items = $service->getBasketItems($basketId);
+        static::assertCount(2, $items->getItems());
+
+        $service->deleteUserBasket();
+
+        static::assertNull($service->getUserBasketId());
     }
 
     private function buildAuthenticatedBasketService(string $email = "admin@wizaplace.com", string $password = "password"): BasketService